+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ## Bug Fixes
 
 - **Crash when changing _[HUD] Layout_ without loading into a map since launching the game**
+- **_Bobbing Weapon Alignment_ misbehaving with DeHackEd-set sprite offsets**
 - **Translucent weapon flashes and hitscan trails not being translucent if _Sprite Translucency_ were disabled**
 - **Sprite shadows not interpolating on moving floors**
 - **Hitscan trails becoming opaque when loading saves**